+++ src/models/mod.rs
@@ -1,4 +1,2 @@
 pub mod common;
-pub mod plane;
-pub mod quaternion;
 pub mod vector_3d;
